requirements.txt
fastapi
uvicorn
yfinance
pandas
pandas_ta
google-genai
python-dotenv
requests
feedparser
pydantic
langchain
langchain-google-genai
